pyproject.toml
[project]
name = "zap-task"
version = "0.1.0"
description = ""
authors = [
    {name = "github-main-user",email = "kolya.skripn@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "psycopg[binary] (>=3.2.9,<4.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "drf-nested-routers (>=0.94.2,<0.95.0)",
    "django-filter (>=25.1,<26.0)",
    "django-fsm-2 (>=4.0.0,<5.0.0)",
    "celery[redis] (>=5.5.3,<6.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-django = "^4.11.1"
coverage = "^7.9.2"

